refactor(contact): flatten entity source structure
