+++ .changeset/flat-contacts-structure.md
@@ -1,0 +1,5 @@
+---
+"@domain/entity-contact": patch
+---
+
+Flatten the Contacts entity source structure.
