+++ domain/entity/contact/src/define.ts
