+++ domain/entity/contact/src/index.ts
@@ -1,5 +1,5 @@
-export * from "./data/schema";
-export * from "./data/types";
-export * from "./data/define";
-export * from "./data/slice";
-export * from "./data/selectors";
+export * from "./schema";
+export * from "./types";
+export * from "./define";
+export * from "./slice";
+export * from "./selectors";
