+++ domain/entity/contact/src/internals.test.ts
