+++ domain/entity/contact/src/internals.ts
